requirements.txt
asgiref==3.6.0
Django==4.1.5
django-catalog==2.1.1
django-classy-tags==4.0.0
django-cors-headers==3.13.0
django-environ==0.9.0
django-filter==22.1
django-js-asset==2.0.0
django-model-mixins==0.1.3
django-mptt==0.14.0
djangorestframework==3.14.0
extdirect.django==0.3
Markdown==3.4.1
psycopg2-binary==2.9.5
pytz==2022.7.1
sqlparse==0.4.3
